pyproject.toml
[project]
name = "scrapes-mcgee"
version = "0.1.0"
description = "LLM-guided web scraper with personality — talk to Scrapes McGee, get targeted content extraction"
authors = [{name = "Merry", email = "merrypranxter@users.noreply.github.com"}]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "google-generativeai>=0.8.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]
js = [
    "playwright>=1.42.0",
]

[project.scripts]
scrapes = "scrapes:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
